requirements.txt
celery==5.4.0
coverage==7.6.1
dj-rest-auth==6.0.0
django==5.1
django-allauth==64.2.0
django-celery-beat==2.7.0
django-constance[database]==2.9.1
django-redis==5.4.0
django-simple-history==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf_spectacular==0.27.2
factory_boy==3.3.1
freezegun==1.5.1
psycopg2-binary==2.9.9
requests==2.32.3
unittest-xml-reporting==3.2.0
